feature/cli-arguments Added command line arguments
* Copied arguments from Config as is which might not make sense after
implementing a PID controller.
